requirements.txt
appdirs==1.4.4
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
elementpath==2.0.4
et-xmlfile==1.0.1
filelock==3.0.12
idna==2.10
jdcal==1.4.1
lxml==4.6.2
numpy==1.19.4
openpyxl==3.0.5
pandas==1.1.4
pipenv==2020.11.15
psycopg2==2.8.6
pycodestyle==2.6.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.1
sh==1.14.1
simplified-scrapy==1.5.159
six==1.15.0
toml==0.10.2
urllib3==1.26.2
virtualenv==20.2.2
virtualenv-clone==0.5.4
xlrd==1.2.0
xmlschema==1.3.1
